BAU: Split driving licence contract tests into two contracts
