+++ .secrets.baseline
@@ -676,90 +676,69 @@
         "line_number": 1351
       }
     ],
-    "lambdas/process-cri-callback/src/test/java/uk/gov/di/ipv/core/processcricallback/pact/drivingLicenceCri/ContractTest.java": [
+    "lambdas/process-cri-callback/src/test/java/uk/gov/di/ipv/core/processcricallback/pact/drivingLicenceCri/CredentialTests.java": [
       {
         "type": "Secret Keyword",
-        "filename": "lambdas/process-cri-callback/src/test/java/uk/gov/di/ipv/core/processcricallback/pact/drivingLicenceCri/ContractTest.java",
+        "filename": "lambdas/process-cri-callback/src/test/java/uk/gov/di/ipv/core/processcricallback/pact/drivingLicenceCri/CredentialTests.java",
         "hashed_secret": "706f9a779d4f10497eb54c0c1a2d3212b54fdb0c",
+        "is_verified": false,
+        "line_number": 64
+      },
+      {
+        "type": "Base64 High Entropy String",
+        "filename": "lambdas/process-cri-callback/src/test/java/uk/gov/di/ipv/core/processcricallback/pact/drivingLicenceCri/CredentialTests.java",
+        "hashed_secret": "4fe0cb4b0c3ce482cb6a43c1931c54502e95ca6f",
+        "is_verified": false,
+        "line_number": 69
+      },
+      {
+        "type": "Base64 High Entropy String",
+        "filename": "lambdas/process-cri-callback/src/test/java/uk/gov/di/ipv/core/processcricallback/pact/drivingLicenceCri/CredentialTests.java",
+        "hashed_secret": "c5752e536a022e8e8c02798c9d785b795ff6983b",
+        "is_verified": false,
+        "line_number": 69
+      },
+      {
+        "type": "Base64 High Entropy String",
+        "filename": "lambdas/process-cri-callback/src/test/java/uk/gov/di/ipv/core/processcricallback/pact/drivingLicenceCri/CredentialTests.java",
+        "hashed_secret": "eada7a0758b629fd871b680e59ca482986eb835b",
+        "is_verified": false,
+        "line_number": 69
+      },
+      {
+        "type": "Base64 High Entropy String",
+        "filename": "lambdas/process-cri-callback/src/test/java/uk/gov/di/ipv/core/processcricallback/pact/drivingLicenceCri/CredentialTests.java",
+        "hashed_secret": "646b4a06350332d91a8ce04fd49101bb4e274f04",
         "is_verified": false,
         "line_number": 73
       },
       {
         "type": "Base64 High Entropy String",
-        "filename": "lambdas/process-cri-callback/src/test/java/uk/gov/di/ipv/core/processcricallback/pact/drivingLicenceCri/ContractTest.java",
-        "hashed_secret": "4fe0cb4b0c3ce482cb6a43c1931c54502e95ca6f",
-        "is_verified": false,
-        "line_number": 78
-      },
-      {
-        "type": "Base64 High Entropy String",
-        "filename": "lambdas/process-cri-callback/src/test/java/uk/gov/di/ipv/core/processcricallback/pact/drivingLicenceCri/ContractTest.java",
-        "hashed_secret": "c5752e536a022e8e8c02798c9d785b795ff6983b",
-        "is_verified": false,
-        "line_number": 78
-      },
-      {
-        "type": "Base64 High Entropy String",
-        "filename": "lambdas/process-cri-callback/src/test/java/uk/gov/di/ipv/core/processcricallback/pact/drivingLicenceCri/ContractTest.java",
-        "hashed_secret": "eada7a0758b629fd871b680e59ca482986eb835b",
-        "is_verified": false,
-        "line_number": 78
-      },
-      {
-        "type": "Base64 High Entropy String",
-        "filename": "lambdas/process-cri-callback/src/test/java/uk/gov/di/ipv/core/processcricallback/pact/drivingLicenceCri/ContractTest.java",
-        "hashed_secret": "646b4a06350332d91a8ce04fd49101bb4e274f04",
-        "is_verified": false,
-        "line_number": 82
-      },
-      {
-        "type": "Base64 High Entropy String",
-        "filename": "lambdas/process-cri-callback/src/test/java/uk/gov/di/ipv/core/processcricallback/pact/drivingLicenceCri/ContractTest.java",
-        "hashed_secret": "28dca06207b1c570f7d1b25ae549d37bdeb93353",
-        "is_verified": false,
-        "line_number": 89
-      },
-      {
-        "type": "Base64 High Entropy String",
-        "filename": "lambdas/process-cri-callback/src/test/java/uk/gov/di/ipv/core/processcricallback/pact/drivingLicenceCri/ContractTest.java",
-        "hashed_secret": "5a2eb5c8850b9e4c2cd896826d871114652ac2f1",
-        "is_verified": false,
-        "line_number": 92
-      },
-      {
-        "type": "Base64 High Entropy String",
-        "filename": "lambdas/process-cri-callback/src/test/java/uk/gov/di/ipv/core/processcricallback/pact/drivingLicenceCri/ContractTest.java",
+        "filename": "lambdas/process-cri-callback/src/test/java/uk/gov/di/ipv/core/processcricallback/pact/drivingLicenceCri/CredentialTests.java",
         "hashed_secret": "4f9fce5708c37ee363ec1627d3fec390a8610b5f",
         "is_verified": false,
-        "line_number": 179
+        "line_number": 163
       },
       {
         "type": "Base64 High Entropy String",
-        "filename": "lambdas/process-cri-callback/src/test/java/uk/gov/di/ipv/core/processcricallback/pact/drivingLicenceCri/ContractTest.java",
+        "filename": "lambdas/process-cri-callback/src/test/java/uk/gov/di/ipv/core/processcricallback/pact/drivingLicenceCri/CredentialTests.java",
         "hashed_secret": "9f9d4c7ca0656e4d1ea4bb9b4932e09435973b82",
         "is_verified": false,
-        "line_number": 256
+        "line_number": 240
       },
       {
         "type": "Base64 High Entropy String",
-        "filename": "lambdas/process-cri-callback/src/test/java/uk/gov/di/ipv/core/processcricallback/pact/drivingLicenceCri/ContractTest.java",
+        "filename": "lambdas/process-cri-callback/src/test/java/uk/gov/di/ipv/core/processcricallback/pact/drivingLicenceCri/CredentialTests.java",
         "hashed_secret": "35bd684504646c60b54288575e917284d043c618",
         "is_verified": false,
-        "line_number": 333
+        "line_number": 317
       },
       {
         "type": "Base64 High Entropy String",
-        "filename": "lambdas/process-cri-callback/src/test/java/uk/gov/di/ipv/core/processcricallback/pact/drivingLicenceCri/ContractTest.java",
+        "filename": "lambdas/process-cri-callback/src/test/java/uk/gov/di/ipv/core/processcricallback/pact/drivingLicenceCri/CredentialTests.java",
         "hashed_secret": "3dd37bc4b5083b9bb330b228bd4a40c5fba200af",
         "is_verified": false,
-        "line_number": 409
-      },
-      {
-        "type": "Base64 High Entropy String",
-        "filename": "lambdas/process-cri-callback/src/test/java/uk/gov/di/ipv/core/processcricallback/pact/drivingLicenceCri/ContractTest.java",
-        "hashed_secret": "6a306e57e40da3db20acfd887b1a6111a2e7b7ce",
-        "is_verified": false,
-        "line_number": 469
+        "line_number": 393
       }
     ],
     "lambdas/process-cri-callback/src/test/java/uk/gov/di/ipv/core/processcricallback/pact/f2fCri/ContractTest.java": [
@@ -1747,5 +1726,5 @@
       }
     ]
   },
-  "generated_at": "2024-02-20T16:37:59Z"
+  "generated_at": "2024-02-22T11:10:36Z"
 }
